# Speed up arithmetic operations involving Proxy subclasses