Arithmetic operations on subclasses of `Proxy` (specifically the C implementation) are slower (roughly three times) due to inexact match in the `PyObject_Instance` call in `Proxy__WRAPPED_REPLACE_OR_RETURN_NULL`.  For simple C-level type compatibility checking `PyObject_TypeCheck` is much faster.  In fact the `PyObject_IsInstance` is probably superfluous with this but I left it anyways.

### Before:

```
In [1]: from lazy_object_proxy import Proxy

In [2]: a = Proxy(int)

In [3]: str(a)
Out[3]: '0'

In [4]: %timeit -n 1000000 -r 100 a + a
1000000 loops, best of 100: 61.1 ns per loop

In [5]: class MyProxy(Proxy): pass

In [6]: b = MyProxy(int)

In [7]: str(b)
Out[7]: '0'

In [8]: %timeit -n 1000000 -r 100 b + b
1000000 loops, best of 100: 171 ns per loop
```

### After:

```
In [1]: from lazy_object_proxy import Proxy

In [2]: a = Proxy(int)

In [3]: str(a)
Out[3]: '0'

In [4]: %timeit -n 1000000 -r 100 a + a
1000000 loops, best of 100: 57.4 ns per loop

In [5]: class MyProxy(Proxy): pass

In [6]: b = MyProxy(int)

In [7]: str(b)
Out[7]: '0'

In [8]: %timeit -n 1000000 -r 100 b + b
1000000 loops, best of 100: 62.1 ns per loop
```