@ionelmc 2017-04-11
Good catch! Now I'm thinking we don't need to do a full `isinstance` check (we don't care about things like https://docs.python.org/3/reference/datamodel.html?highlight=__instancecheck__#class.__instancecheck__) since a "fake instance" won't have the right memory layout anyway ...