@embray 2017-04-11
I was thinking that too over lunch (it was a very lonely lunch :)