@embray 2017-04-11
Would you like me to update it to just remove the full `PyObject_IsInstance`?