@ionelmc 2017-04-11
Also, update CHANGELOG/AUTHORS.rst.