@embray 2017-04-11
Done--I agree it makes sense to submit the change upstream to wrapt as well.