@ionelmc 2017-05-04
Published `1.3.0` on PyPI.